problem with owner

+++ controllers/contacts/deleteContact.js
@@ -4,6 +4,16 @@ async function deleteContact(req, res, next) {
   const { contactId } = req.params;
 
   try {
+    const contact = Contact.findById(contactId);
+
+    if (contact === null) {
+      return res.status(404).json({ message: "Not found" });
+    }
+
+    if (contact.owner.toString() !== req.user.id) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
+
     const result = await Contact.findByIdAndDelete(contactId);
 
     if (result === null) {
